requirements.txt
streamlit
replicate
pillow
transformers
BERTSimilarity==0.1

